Basic program (debug)
This program can calibrate the cameras, display the two views in RGB or
YCrCb, save a sequence in AVI or YUV format, and show the two view on
top of each other so that the user can see the scene in 3D with
anaglyphic glasses. This is a debug version : you can see the time each
function takes to execute.