Major commit
Implementing a client-server application -> new project :
3DWebcamServer.
Resizing of the frame for faster treatment.
Adding a chat.
Cleaning the code to separate all tasks in different classes
(VideoHandler, VideoThread).
Buttons used for tests are in another window.
Implementing the possibility to encode 15 frames anytime. The encoded
file (*.264) can be sent via the chat.